+++ src/main/java/com/protonmail/landrevillejf/cognos/categories/api/entity/dto/ErrorDTO.java
@@ -1,0 +1,14 @@
+package com.protonmail.landrevillejf.cognos.categories.api.entity.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class ErrorDTO {
+
+    private String field;
+    private String errorMessage;
+}