+++ src/main/java/com/protonmail/landrevillejf/cognos/categories/api/entity/dto/FileDTO.java
@@ -1,0 +1,15 @@
+package com.protonmail.landrevillejf.cognos.categories.api.entity.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class FileDTO {
+
+    private String fileContent;
+    private String fileName;
+
+}