+++ src/main/java/com/protonmail/landrevillejf/cognos/categories/api/payload/package-info.java
@@ -1,0 +1,1 @@
+package com.protonmail.landrevillejf.cognos.categories.api.payload;